# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Внутренний API дневника тренировок",
    "description": "API для доверенных сервисов. X-Api-Key проверяет вызывающий сервис; X-Telegram-User-Id выбирает пользователя. Не передавайте ключ клиентским приложениям: он позволяет действовать от имени любого пользователя. Вес в кг, расстояние в км, длительность в секундах.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "apiKey": [],
      "telegramIdentity": []
    }
  ],
  "paths": {
...
    "/api/training/settings": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "preferences",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "training-controller"
        ],
        "operationId": "preferences_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Preferences"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/sets/{setId}": {
      "put": {
        "tags": [
          "training-controller"
        ],
        "operationId": "edit_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "setId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workouts/import": {
      "post": {
        "tags": [
          "workout-controller"
        ],
        "operationId": "importCsv",
        "parameters": [
          {
            "name": "replaceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "allowDuplicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/template": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/rest": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "rest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/pause": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "pause",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pause"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/finish-at": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "finish",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Finish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/exercises": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "change",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Change"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/schedule": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "schedule",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "schedule_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Schedule"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/history": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "history",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/history/import": {
      "post": {
        "tags": [
          "training-controller"
        ],
        "operationId": "importHistory",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "user-controller"
        ],
        "operationId": "settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions/{id}/position": {
      "patch": {
        "tags": [
          "session-controller"
        ],
        "operationId": "position",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/suggestions/{id}": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "suggest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestion"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/sessions/{id}/comparison": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "comparison",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comparison"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/training/history/export": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "exportHistory",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/csv;charset=UTF-8": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "history_1",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/schedule/{id}": {
      "delete": {
        "tags": [
          "training-controller"
        ],
        "operationId": "cancel_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/sets/{setId}": {
      "delete": {
        "tags": [
          "session-controller"
        ],
        "operationId": "undo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "setId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TemplateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          }
        }
      },
...
      "TemplateView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateItem"
            }
          }
        }
      },
      "Preferences": {
        "type": "object",
        "properties": {
          "weeklyGoal": {
            "type": "integer",
            "format": "int32"
          },
          "restNotifications": {
            "type": "boolean"
          },
          "weightStep": {
            "type": "number"
          },
          "repMin": {
            "type": "integer",
            "format": "int32"
          },
          "repMax": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SetInput": {
        "type": "object",
        "properties": {
          "sessionExerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "weight": {
            "type": "number"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "distance": {
            "type": "number"
          },
          "rpe": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          },
          "warmup": {
            "type": "boolean"
          }
        }
      },
      "ExerciseView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "metricType": {
            "type": "string",
            "enum": [
              "STRENGTH",
              "BODYWEIGHT",
              "CARDIO",
              "TIMED"
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "targetSets": {
            "type": "integer",
            "format": "int32"
          },
          "targetReps": {
            "type": "integer",
            "format": "int32"
          },
          "targetWeight": {
            "type": "number"
          },
          "restSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetView"
            }
          },
          "targetDuration": {
            "type": "integer",
            "format": "int32"
          },
          "targetDistance": {
            "type": "number"
          },
          "plan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetPlan"
            }
          },
          "skipped": {
            "type": "boolean"
          }
        }
      },
      "SessionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentPosition": {
            "type": "integer",
            "format": "int32"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseView"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "pausedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pausedSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "restUntil": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Summary": {
        "type": "object",
        "properties": {
          "durationSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "exercises": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "repetitions": {
            "type": "integer",
            "format": "int64"
          },
          "volume": {
            "type": "number"
          }
        }
      },
      "Rest": {
        "type": "object",
        "properties": {
          "seconds": {
            "type": "integer",
            "format": "int32"
          },
          "extend": {
            "type": "boolean"
          }
        }
      },
      "Pause": {
        "type": "object",
        "properties": {
          "paused": {
            "type": "boolean"
          }
        }
      },
      "Finish": {
        "type": "object",
        "properties": {
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Change": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "skipped": {
            "type": "boolean"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "localStart": {
            "type": "string",
            "format": "date-time"
          },
          "repeatWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "reminders": {
            "type": "boolean"
          }
        }
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "templateId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "localStart": {
            "type": "string",
            "format": "date-time"
          },
          "repeatWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "reminders": {
            "type": "boolean"
          }
        }
      },
      "HistoricalExercise": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int64"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetInput"
            }
          }
        }
      },
      "HistoricalInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoricalExercise"
            }
          },
          "pausedSeconds": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Registration": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          }
        }
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "telegramUserId": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
...
      "Position": {
        "type": "object",
        "properties": {
          "position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Settings": {
        "type": "object",
        "properties": {
          "timezone": {
            "type": "string"
          },
          "language": {
            "type": "string"
          }
        }
      },
      "Suggestion": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/TemplateInput"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "Comparison": {
        "type": "object",
        "properties": {
          "exercise": {
            "type": "string"
          },
          "planned": {
            "type": "integer",
            "format": "int32"
          },
          "completed": {
            "type": "integer",
            "format": "int32"
          },
          "extra": {
            "type": "integer",
            "format": "int32"
          },
          "skipped": {
            "type": "boolean"
          },
          "previousVolume": {
            "type": "number"
          },
          "currentVolume": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header"
      },
      "telegramIdentity": {
        "type": "apiKey",
        "name": "X-Telegram-User-Id",
        "in": "header"
      }
    }
  }
}